pyproject.toml
[tool.poetry]
name = "workspace"
version = "0.1.0"
description = ""
authors = ["Jose Valencia <jose.valencia@ctsturismo.cl>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
psycopg = "^3.1.9"
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
django-filter = "^23.2"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^4.1.0"
django-environ = "^0.11.2"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
